改了点

+++ Hardware/Motor.h
@@ -4,7 +4,7 @@
 #include "stm32f10x.h"
 
 void Motor_Init(void);
-void Motor1_SetPWM(int16_t PWM);
-void Motor2_SetPWM(int16_t PWM);
+void Motor1_SetPWM(int8_t PWM);
+void Motor2_SetPWM(int8_t PWM);
 
 #endif
